ISU-87 Update role path edit
The role path set in OSA has been adjusted to
be more absolute, so our modification of it
needs to be adjusted to edit it properly.

Instead of replacing the whole string as we
did previously, we now just insert the RPC-O
role paths in between the two OSA paths.