AR-206 Make chroot only member of role group
When building the container artifacts, some
install tasks make use of inventory membership
checks. This is usually to check for the first
member of the <role name>_all group.

This patch ensures that the container artifact
builds include the chroot as the only member
of the <role name>_all group.